requirements.txt
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
nltk>=3.8.1
joblib
numpy
scipy
newspaper3k
lxml_html_clean
